REF: split TabDockButton into its own file to serve as a good standalone example
